+++ package.json
@@ -2,7 +2,7 @@
   "name": "@walk8243/logger",
   "version": "3.0.0",
   "description": "log4jsの設定にデフォルトを設ける",
-  "main": "lib/logger.js",
+  "main": "logger.js",
   "scripts": {
     "build": "yarn run tsc --sourceMap false --removeComments true",
     "build-debug": "yarn run tsc --build",
